requirements.txt
fastapi
uvicorn
slowapi
python-dotenv
groq
pydantic
chromadb
sentence-transformers
pypdf
